Added an option to make masters un/schedulable
As the master nodes were set to schedulable by default, I added an option that allows/disallows this. It adds two tasks before creating the ingition files in `create-ignitions.yml`.
An entry was added in the readme file with the optional configuration and a default value (set to false) was added to the defaults/main.yml file.

Done and tested on a CentOS 8.2 machine.